set up the small changes needed to upgrade to mujoco 3.5.0 removed references to package visualscripting, which was not needed

+++ modular-agents/mujoco/MujocoModularAgents/Runtime/RL Signals/GRFObservationSource.cs
@@ -46,7 +46,7 @@ namespace ModularAgents
             }
             else
             {
-                includedIds = Enumerable.Range(0, MjScene.Instance.Model->ngeom);
+                includedIds = Enumerable.Range(0, (int) MjScene.Instance.Model->ngeom);
             }
 
             List<Vector3> pos = new List<Vector3>();
